+++ pyocd/flash/loader.py
@@ -14,353 +14,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import print_function
-import os
 import logging
-import itertools
-from intelhex import IntelHex
-from enum import Enum
-import six
-import errno
 
-from .flash_builder import (FlashBuilder, get_page_count, get_sector_count)
-from ..core.memory_map import MemoryType
+from .builder import (FlashBuilder, get_page_count, get_sector_count)
 from ..core import exceptions
 from ..utility.progress import print_progress
-from elftools.elf.elffile import ELFFile
-from ..utility.compatibility import FileNotFoundError_
 
 LOG = logging.getLogger(__name__)
-
-def ranges(i):
-    """!
-    Accepts a sorted list of byte addresses. Breaks the addresses into contiguous ranges.
-    Yields 2-tuples of the start and end address for each contiguous range.
-    
-    For instance, the input [0, 1, 2, 3, 32, 33, 34, 35] will yield the following 2-tuples:
-    (0, 3) and (32, 35).
-    """
-    for a, b in itertools.groupby(enumerate(i), lambda x: x[1] - x[0]):
-        b = list(b)
-        yield b[0][1], b[-1][1]
-
-class FileProgrammer(object):
-    """! @brief Class to manage programming a file in any supported format with many options.
-    
-    Most specifically, this class implements the behaviour provided by the command-line flash
-    programming tool. The code in this class simply extracts data from the given file, potentially
-    respecting format-specific options such as the base address for binary files. Then the heavy
-    lifting of flash programming is handled by FlashLoader, and beneath that, FlashBuilder.
-    
-    Support file formats are:
-    - Binary (.bin)
-    - Intel Hex (.hex)
-    - ELF (.elf or .axf)
-    """
-    def __init__(self, session, progress=None, chip_erase=None, smart_flash=None,
-        trust_crc=None, keep_unwritten=None):
-        """! @brief Constructor.
-        
-        @param self
-        @param session The session object.
-        @param progress A progress report handler as a callable that takes a percentage completed.
-            If not set or None, a default progress handler will be used unless the session option
-            'hide_programming_progress' is set to True, in which case progress will be disabled.
-        @param chip_erase Sets whether to use chip erase or sector erase. The value must be one of
-            "auto", "sector", or "chip". "auto" means the fastest erase method should be used.
-        @param smart_flash If set to True, the programmer will attempt to not program pages whose
-            contents are not going to change by scanning target flash memory. A value of False will
-            force all pages to be erased and programmed.
-        @param trust_crc Boolean indicating whether to use only the sector CRC32 to decide whether a
-            sector already contains the data to be programmed. Use with caution, as CRC32 may return
-            the same value for different content.
-        @param keep_unwritten Depending on the sector versus page size and the amount of data
-            written, there may be ranges of flash that would be erased but not written with new
-            data. This parameter sets whether the existing contents of those unwritten ranges will
-            be read from memory and restored while programming.
-        """
-        self._session = session
-        self._chip_erase = chip_erase
-        self._smart_flash = smart_flash
-        self._trust_crc = trust_crc
-        self._keep_unwritten = keep_unwritten
-        self._progress = progress
-        
-        self._format_handlers = {
-            'axf': self._program_elf,
-            'bin': self._program_bin,
-            'elf': self._program_elf,
-            'hex': self._program_hex,
-            }
-    
-    def program(self, file_or_path, file_format=None, **kwargs):
-        """! @brief Program a file into flash.
-        
-        @param self
-        @param file_or_path Either a string that is a path to a file, or a file-like object.
-        @param file_format Optional file format name, one of "bin", "hex", "elf", "axf". If not provided,
-            the file's extension will be used. If a file object is passed for _file_or_path_ then
-            this parameter must be used to set the format.
-        @param kwargs Optional keyword arguments for format-specific parameters.
-        
-        The only current format-specific keyword parameters are for the binary format:
-        - `base_address`: Memory address at which to program the binary data. If not set, the base
-            of the boot memory will be used.
-        - `skip`: Number of bytes to skip at the start of the binary file. Does not affect the
-            base address.
-        
-        @exception FileNotFoundError Provided file_or_path string does not reference a file.
-        @exception ValueError Invalid argument value, for instance providing a file object but
-            not setting file_format.
-        """
-        isPath = isinstance(file_or_path, six.string_types)
-        
-        # Check for valid path first.
-        if isPath and not os.path.isfile(file_or_path):
-            raise FileNotFoundError_(errno.ENOENT, "No such file: '{}'".format(file_or_path))
-        
-        # If no format provided, use the file's extension.
-        if not file_format:
-            if isPath:
-                # Extract the extension from the path.
-                file_format = os.path.splitext(file_or_path)[1][1:]
-                
-                # Explicitly check for no extension.
-                if file_format == '':
-                    raise ValueError("file path '{}' does not have an extension and "
-                                        "no format is set".format(file_or_path))
-            else:
-                raise ValueError("file object provided but no format is set")
-        
-        # Check the format is one we understand.
-        if file_format not in self._format_handlers:
-            raise ValueError("unknown file format '%s'" % file_format)
-            
-        self._loader = FlashLoader(self._session,
-                                    progress=self._progress,
-                                    chip_erase=self._chip_erase,
-                                    smart_flash=self._smart_flash,
-                                    trust_crc=self._trust_crc,
-                                    keep_unwritten=self._keep_unwritten)
-        
-        file_obj = None
-        try:
-            # Open the file if a path was provided.
-            if isPath:
-                mode = 'rb'
-                if file_format == 'hex':
-                    # hex file must be read as plain text file
-                    mode = 'r'
-                file_obj = open(file_or_path, mode)
-            else:
-                file_obj = file_or_path
-
-            # Pass to the format-specific programmer.
-            self._format_handlers[file_format](file_obj, **kwargs)
-            self._loader.commit()
-        finally:
-            if isPath and file_obj is not None:
-                file_obj.close()
-
-    def _program_bin(self, file_obj, **kwargs):
-        """! @brief Binary file format loader"""
-        # If no base address is specified use the start of the boot memory.
-        address = kwargs.get('base_address', None)
-        if address is None:
-            boot_memory = self._session.target.memory_map.get_boot_memory()
-            if boot_memory is None:
-                raise exceptions.TargetSupportError("No boot memory is defined for this device")
-            address = boot_memory.start
-        
-        file_obj.seek(kwargs.get('skip', 0), os.SEEK_SET)
-        data = list(bytearray(file_obj.read()))
-        
-        self._loader.add_data(address, data)
-
-    def _program_hex(self, file_obj, **kwargs):
-        """! Intel hex file format loader"""
-        hexfile = IntelHex(file_obj)
-        addresses = hexfile.addresses()
-        addresses.sort()
-
-        data_list = list(ranges(addresses))
-        for start, end in data_list:
-            size = end - start + 1
-            data = list(hexfile.tobinarray(start=start, size=size))
-            # Ignore invalid addresses for HEX files only
-            # Binary files (obviously) don't contain addresses
-            # For ELF files, any metadata that's not part of the application code 
-            # will be held in a section that doesn't have the SHF_WRITE flag set
-            try:
-                self._loader.add_data(start, data)
-            except ValueError as e:
-                LOG.warning("Failed to add data chunk: %s", e)
-
-    def _program_elf(self, file_obj, **kwargs):
-        elf = ELFFile(file_obj)
-        for segment in elf.iter_segments():
-            addr = segment['p_paddr']
-            if segment.header.p_type == 'PT_LOAD' and segment.header.p_filesz != 0:
-                data = bytearray(segment.data())
-                LOG.debug("Writing segment LMA:0x%08x, VMA:0x%08x, size %d", addr, 
-                          segment['p_vaddr'], segment.header.p_filesz)
-                try:
-                    self._loader.add_data(addr, data)
-                except ValueError as e:
-                    LOG.warning("Failed to add data chunk: %s", e)
-            else:
-                LOG.debug("Skipping segment LMA:0x%08x, VMA:0x%08x, size %d", addr,
-                          segment['p_vaddr'], segment.header.p_filesz)
-                        
-
-class FlashEraser(object):
-    """! @brief Class that manages high level flash erasing.
-    
-    Can erase a target in one of three modes:
-    - chip erase: Erase all flash on the target.
-    - mass erase: Also erase all flash on the target. However, on some targets, a mass erase has
-        special properties such as unlocking security or erasing additional configuration regions
-        that are not erased by a chip erase. If a target does not have a special mass erase, then
-        it simply reverts to a chip erase.
-    - sector erase: One or more sectors are erased.
-    """
-    class Mode(Enum):
-        MASS = 1
-        CHIP = 2
-        SECTOR = 3
-    
-    def __init__(self, session, mode):
-        """! @brief Constructor.
-        
-        @param self
-        @param session The session instance.
-        @param mode One of the FlashEraser.Mode enums to select mass, chip, or sector erase.
-        """
-        self._session = session
-        self._mode = mode
-    
-    def erase(self, addresses=None):
-        """! @brief Perform the type of erase operation selected when the object was created.
-        
-        For sector erase mode, an iterable of sector addresses specifications must be provided via
-        the _addresses_ parameter. The address iterable elements can be either strings, tuples,
-        or integers. Tuples must have two elements, the start and end addresses of a range to erase.
-        Integers are simply an address within the single page to erase.
-        
-        String address specifications may be in one of three formats: "<address>", "<start>-<end>",
-        or "<start>+<length>". Each field denoted by angled brackets is an integer literal in
-        either decimal or hex notation.
-        
-        Examples:
-        - "0x1000" - erase the one sector at 0x1000
-        - "0x1000-0x4fff" - erase sectors from 0x1000 up to but not including 0x5000
-        - "0x8000+0x800" - erase sectors starting at 0x8000 through 0x87ff
-        
-        @param self
-        @param addresses List of addresses or address ranges of the sectors to erase.
-        """
-        if self._mode == self.Mode.MASS:
-            self._mass_erase()
-        elif self._mode == self.Mode.CHIP:
-            self._chip_erase()
-        elif self._mode == self.Mode.SECTOR and addresses:
-            self._sector_erase(addresses)
-        else:
-            LOG.warning("No operation performed")
-    
-    def _mass_erase(self):
-        LOG.info("Mass erasing device...")
-        if self._session.target.mass_erase():
-            LOG.info("Successfully erased.")
-        else:
-            LOG.error("Mass erase failed.")
-    
-    def _chip_erase(self):
-        LOG.info("Erasing chip...")
-        # Erase all flash regions. This may be overkill if either each region's algo erases
-        # all regions on the chip. But there's no current way to know whether this will happen,
-        # so prefer to be certain.
-        for region in self._session.target.memory_map.iter_matching_regions(type=MemoryType.FLASH):
-            if region.flash is not None:
-                if region.flash.is_erase_all_supported:
-                    region.flash.init(region.flash.Operation.ERASE)
-                    region.flash.erase_all()
-                    region.flash.cleanup()
-                else:
-                    self._sector_erase((region.start, region.end))
-        LOG.info("Done")
-    
-    def _sector_erase(self, addresses):
-        flash = None
-        currentRegion = None
-
-        for spec in addresses:
-            # Convert the spec into a start and end address.
-            sector_addr, end_addr = self._convert_spec(spec)
-            
-            while sector_addr < end_addr:
-                # Look up the flash memory region for the current address.
-                region = self._session.target.memory_map.get_region_for_address(sector_addr)
-                if region is None:
-                    LOG.warning("address 0x%08x is not within a memory region", sector_addr)
-                    break
-                if not region.is_flash:
-                    LOG.warning("address 0x%08x is not in flash", sector_addr)
-                    break
-            
-                # Handle switching regions.
-                if region is not currentRegion:
-                    # Clean up previous flash.
-                    if flash is not None:
-                        flash.cleanup()
-                
-                    currentRegion = region
-                    flash = region.flash
-                    flash.init(flash.Operation.ERASE)
-        
-                # Get sector info for the current address.
-                sector_info = flash.get_sector_info(sector_addr)
-                if not sector_info:
-                    # Should not fail to get sector info within a flash region.
-                    raise exceptions.InternalError("sector address 0x%08x within flash region '%s' is invalid" % (sector_addr, region.name))
-                
-                # Align first page address.
-                delta = sector_addr % sector_info.size
-                if delta:
-                    LOG.warning("sector address 0x%08x is unaligned", sector_addr)
-                    sector_addr -= delta
-        
-                # Erase this page.
-                LOG.info("Erasing sector 0x%08x (%d bytes)", sector_addr, sector_info.size)
-                flash.erase_sector(sector_addr)
-                
-                sector_addr += sector_info.size
-
-        if flash is not None:
-            flash.cleanup()
-
-    def _convert_spec(self, spec):
-        if isinstance(spec, six.string_types):
-            # Convert spec from string to range.
-            if '-' in spec:
-                a, b = spec.split('-')
-                page_addr = int(a, base=0)
-                end_addr = int(b, base=0)
-            elif '+' in spec:
-                a, b = spec.split('+')
-                page_addr = int(a, base=0)
-                length = int(b, base=0)
-                end_addr = page_addr + length
-            else:
-                page_addr = int(spec, base=0)
-                end_addr = page_addr + 1
-        elif isinstance(spec, tuple):
-            page_addr = spec[0]
-            end_addr = spec[1]
-        else:
-            page_addr = spec
-            end_addr = page_addr + 1
-        return page_addr, end_addr
 
 class FlashLoader(object):
     """! @brief Handles high level programming of raw binary data to flash.
