+++ test/basic_test.py
@@ -25,7 +25,7 @@ import logging
 
 from pyocd.core.helpers import ConnectHelper
 from pyocd.core.memory_map import MemoryType
-from pyocd.flash.loader import FileProgrammer
+from pyocd.flash.file_programmer import FileProgrammer
 from pyocd.utility.conversion import float32_to_u32
 
 from test_util import (
